dirCreator {shinyFiles}	R Documentation

Create a function that creates a new directory

Description

This function returns a function that can be used to create new directories based on the information returned from the client. The returned function takes the name, path and root of the directory to create and parses it into a platform compliant format and then uses dir.create to create it. The returned function returns TRUE if the directory was created successfully and FALSE otherwise.

Usage

dirCreator(roots, ...)

Arguments

roots
A named vector of absolute filepaths or a function returning a named vector of absolute filepaths (the latter is useful if the volumes should adapt to changes in the filesystem).

...
Currently unused

Value

A function that creates directories based on the information returned by the client.

[Package shinyFiles version 0.7.3 Index]