formatFiletype {shinyFiles}	R Documentation

Formats the value of the filetype argument

Description

This function is intended to format the filetype argument of shinySaveButton() into a json string representation, so that it can be attached to the button.

Usage

formatFiletype(filetype)

Arguments

filetype
A named list of file extensions or NULL or NA

Value

A string describing the input value in json format

[Package shinyFiles version 0.7.3 Index]