shinyFiles-package {shinyFiles}	R Documentation

A Server-Side File System Viewer for Shiny

Description

Provides functionality for client-side navigation of the server side file system in shiny apps. In case the app is running locally this gives the user direct access to the file system without the need to "download" files to a temporary location. Both file and folder selection as well as file saving is available.

Author(s)

Maintainer: Thomas Lin Pedersen thomasp85@gmail.com

Authors:

Vincent Nijs

Thomas Schaffner

Eric Nantz

See Also

Useful links:

https://github.com/thomasp85/shinyFiles

Report bugs at https://github.com/thomasp85/shinyFiles/issues

[Package shinyFiles version 0.7.3 Index]