updateChildren {shinyFiles}	R Documentation

Update the children element to reflect current state

Description

This function create new entries for new folders and remove entries for no longer existing folders, while keeping the state of transient folders in the children element of the tree structure. The function does not recurse into the folders, but merely creates a shell that traverseDirs can take as input.

Usage

updateChildren(oldChildren, currentChildren)

Arguments

oldChildren
A list of children folders from the parent$children element of tree in traverseDirs()

currentChildren
A vector of names of the folders that are currently present in the parent of oldChildren

Value

An updated list equal in format to oldChildren

[Package shinyFiles version 0.7.3 Index]