getVolumes {shinyFiles}	R Documentation

Get a list of available volumes

Description

This function is intended as an input to the roots parameter in fileGetter() and shinyFileChoose(). It returns a function that returns a named vector of available volumes on the system. This construction makes it dynamic so that a shinyFiles instance reflects new volumes as they get added (e.g. usb drives). The function takes a single argument giving names of volumes the developer wants removed from the return value.

Usage

getVolumes(exclude)

Arguments

exclude
A vector of volume names to be excluded from the return value

Details

The function is OS specific and looks for volumes/drives in different places depending on the system on which shiny is running.

Windows

Returns all drives mapped to a letter

Mac OSX

Looks in /Volumes/ and lists the directories therein

Linux

Returns the system root

If the function does not recognize the system under which it is running it will throw an error

Value

A function returning a named vector of available volumes

[Package shinyFiles version 0.7.3 Index]